Makefile with option between openMP and CUDA
