+++ cpu.txt
@@ -1,0 +1,12 @@
+track in pos: -12.806847, -7.723825, 38.130142 
+track in cov: 6.29e-07, 7.53e-07, 9.63e-08 
+hit in pos: -20.782465 -12.241503 57.806763 
+produce nevts=1000 ntrks=9600 smearing by=0.100000 
+done preparing!
+done ntracks=9600000 tot time=0.094000 (s) time/trk=9.791667e-09 (s)
+track x avg=-18.249071 std/avg=0.156725
+track y avg=-11.098969 std/avg=0.207473
+track z avg=60.430508 std/avg=0.104293
+track dx/x avg=-0.123542 std=0.204652
+track dy/y avg=-0.131736 std=0.314768
+track dz/z avg=0.000000 std=0.000000
